requirements.txt
google-generativeai 
langchain-google-genai 
streamlit 
pillow